backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-multipart==0.0.7
aiofiles==23.1.0
openai==1.99.5
python-dotenv==1.0.1
